# docs/swagger.yaml
basePath: development
definitions:
...
  request.AssetCodeRequest:
    properties:
      brandName:
        type: string
      carCondition:
        type: string
      manufacturedYear:
        type: string
      modelName:
        type: string
      variantName:
        type: string
    type: object
  request.PackageRequest:
    properties:
      brand:
        type: string
      carCondition:
        type: string
      city:
        type: string
      model:
        type: string
      packageName:
        type: string
      province:
        type: string
      variant:
        type: string
    type: object
...
  response.AssetCodeResponse:
    properties:
      originalAssetCode:
        type: string
      targetAssetCode:
        type: string
    type: object
  response.BrandsResponse:
    properties:
      data: {}
      message:
        type: string
      recordCount:
        type: integer
      status:
        type: boolean
    type: object
  response.CitiesResponse:
    properties:
      current_total_item:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: boolean
      total_all_item:
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      error_id:
        type: integer
      message:
        additionalProperties:
          type: string
        type: object
    type: object
...
  response.InsuranceResponse:
    properties:
      additionalRates: {}
      basicInsuranceRates: {}
      loadingRate: {}
      palprate: {}
      tjhrates: {}
    type: object
  response.ModelsResponse:
    properties:
      data: {}
      message:
        type: string
      recordCount:
        type: integer
      status:
        type: boolean
    type: object
...
  response.ProvincesResponse:
    properties:
      current_total_item:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: boolean
      total_all_item:
        type: integer
    type: object
...
  response.VariantsResponse:
    properties:
      data:
        items:
          properties:
            name:
              type: string
            product_model_id:
              type: integer
          type: object
        type: array
      message:
        type: string
      recordCount:
        type: integer
      status:
        type: boolean
    type: object
...
host: xuu77ziiri.execute-api.us-east-2.amazonaws.com/
info:
  contact: {}
  description: MMKSI Middleware API Documentation
  title: MMKSI Middleware API Documentation
  version: "1.0"
paths:
...
  /dsf/metadata/additionalInsurance:
    get:
      description: Get Additional Insurance
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              properties:
                name:
                  type: string
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Additional Insurance
      tags:
      - Metadata
  /dsf/metadata/assetCode:
    post:
      description: Get Asset Code
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Application Name
        in: header
        name: applicationName
        required: true
        type: string
      - description: Asset Code Request
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/request.AssetCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AssetCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Asset Code
      tags:
      - Metadata
  /dsf/metadata/branchID:
    get:
      description: Get BranchID
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              properties:
                branchCode:
                  type: string
                branchName:
                  type: string
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get BranchID
      tags:
      - Metadata
  /dsf/metadata/brands:
    get:
      description: Get Brands
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - in: query
        name: keyword
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BrandsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Brands
      tags:
      - Metadata
  /dsf/metadata/carConditions:
    get:
      description: Get Car Condition
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              properties:
                description:
                  type: string
                name:
                  type: string
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Car Condition
      tags:
      - Metadata
  /dsf/metadata/cities:
    get:
      description: Get Cities
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: provinceCode
        type: string
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CitiesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Cities
      tags:
      - Metadata
  /dsf/metadata/insuranceTypes:
    get:
      description: Get Insurance Types
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              properties:
                description:
                  type: string
                name:
                  type: string
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Insurance Types
      tags:
      - Metadata
  /dsf/metadata/insurances:
    get:
      description: Get Insurances
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Dsf Branch ID
        in: query
        name: dsfBranchId
        required: true
        type: string
      - description: Vehicle Category
        in: query
        name: VehicleCategory
        required: true
        type: string
      - description: Insurance Type Code
        in: query
        name: InsuranceTypeCode
        required: true
        type: string
      - description: Car Condition
        in: query
        name: Condition
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.InsuranceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Insurances
      tags:
      - Metadata
  /dsf/metadata/models:
    get:
      description: Get Models
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - in: query
        name: brand
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ModelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Models
      tags:
      - Metadata
  /dsf/metadata/packageNames:
    get:
      description: Get Package Names
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Application Name
        in: header
        name: applicationName
        required: true
        type: string
      - description: Asset Code
        in: header
        name: assetCode
        required: true
        type: string
      - description: Branch Code
        in: header
        name: branchCode
        required: true
        type: string
      - in: query
        name: carCondition
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Package Names
      tags:
      - Metadata
  /dsf/metadata/packages:
    post:
      description: Get Packages
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Application Name
        in: header
        name: applicationName
        required: true
        type: string
      - description: Package
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/request.PackageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Packages
      tags:
      - Metadata
  /dsf/metadata/paymentTypes:
    get:
      description: Get Payment Types
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              properties:
                description:
                  type: string
                name:
                  type: string
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Payment Types
      tags:
      - Metadata
  /dsf/metadata/province:
    get:
      description: Get Province
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProvincesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Province
      tags:
      - Metadata
  /dsf/metadata/variants:
    get:
      description: Get Variants
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      - description: Application Name
        in: header
        name: applicationName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.VariantsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Variants
      tags:
      - Metadata
  /dsf/metadata/vehicleCategory:
    get:
      description: Get Vehicle Category
      parameters:
      - description: Authentication
        in: header
        name: Authentication
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              properties:
                description:
                  type: string
                name:
                  type: string
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Vehicle Category
      tags:
      - Metadata
...
schemes:
- https
- http
swagger: "2.0"